+++ grunt/config/release.js
@@ -1,6 +1,0 @@
-module.exports = {
-	options: {
-		stderr: false,
-		tagName: 'v<%= version %>', //default: '<%= version %>'
-	}
-};