+++ grunt/config/shell.js
@@ -1,8 +1,0 @@
-module.exports = {
-	options: {
-		stderr: false
-	},
-	test: {
-		command: 'npm test'
-	}
-};