+++ grunt/config/todo.js
@@ -1,5 +1,0 @@
-module.exports = {
-	todo: {
-		src: [ '<%= vars.paths.js %>' ]
-	}
-};