+++ grunt/config/watch.js
@@ -1,6 +1,0 @@
-module.exports = {
-	js: {
-		files: [ '<%= vars.paths.js %>' ],
-		tasks: [ 'js_on_watch' ]
-	}
-};